api/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
astroid==3.3.4
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
azure-core==1.30.2
azure-storage-blob==12.20.0
Babel==2.15.0
bandit==1.7.10
beautifulsoup4==4.12.3
bitarray==2.9.2
bs4==0.0.2
build==1.0.3
CacheControl==0.13.1
certifi==2024.2.2
cffi==1.17.1
channels==4.1.0
charset-normalizer==3.3.2
ckzg==1.0.2
cleo==2.1.0
click==8.1.7
colorama==0.4.6
constantly==23.10.4
crashtest==0.4.1
cryptography==42.0.8
cytoolz==0.12.3
daphne==4.1.2
dill==0.3.9
distlib==0.3.8
Django==5.0.4
django-cors-headers==4.3.1
django-environ==0.11.2
django-map-widgets==0.4.2
django-static-precompiler==2.4
django-storages==1.14.3
django-stubs==5.0.4
django-stubs-ext==5.0.4
djangorestframework==3.15.1
drfpasswordless==1.5.9
dulwich==0.21.7
eth-account==0.11.2
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==1.0.1
eth-typing==4.2.3
eth-utils==4.1.1
eth_abi==5.1.0
fastjsonschema==2.19.1
filelock==3.13.1
frozenlist==1.4.1
gunicorn==22.0.0
h11==0.14.0
h2==4.1.0
hexbytes==0.3.1
hpack==4.0.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.6
importlib-metadata==7.0.1
incremental==22.10.0
installer==0.7.0
isodate==0.6.1
isort==5.13.2
jaraco.classes==3.3.1
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keyring==24.3.0
libsass==0.23.0
lru-dict==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.2.0
msgpack==1.1.0
multidict==6.0.5
mypy==1.11.2
mypy-extensions==1.0.0
numpy==1.26.4
outcome==1.3.0.post0
packaging==23.2
parsimonious==0.10.0
pbr==6.1.0
pexpect==4.9.0
pillow==10.4.0
pipenv==2023.12.1
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
priority==1.3.0
protobuf==4.25.3
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pur==7.3.1
py-solc-x==2.0.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
Pygments==2.18.0
PyJWT==2.8.0
pylint==3.3.1
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pyOpenSSL==24.1.0
pypng==0.20220715.0
pyproject_hooks==1.0.0
PySocks==1.7.1
python-dotenv==1.0.1
pytz==2024.1
pyunormalize==16.0.0
pywin32-ctypes==0.2.2
PyYAML==6.0.2
qrcode==7.4.2
rapidfuzz==3.6.1
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.9.1
rlp==4.0.1
rpds-py==0.22.3
selenium==4.17.2
service-identity==24.1.0
setuptools==69.5.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
solc-select==1.0.4
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
stevedore==5.3.0
toml==0.10.2
tomlkit==0.12.3
toolz==0.12.1
trio==0.24.0
trio-websocket==0.11.1
trove-classifiers==2024.1.31
Twisted==24.3.0
txaio==23.1.1
types-PyYAML==6.0.12.20240808
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.0
virtualenv==20.25.1
web3==6.19.0
websockets==12.0
wsproto==1.2.0
xattr==0.10.1
yarl==1.9.4
zipp==3.17.0
zope.interface==6.4.post2
whitenoise==6.8.2
djangorestframework-simplejwt==5.3.1
